// frontend/src/api/entityModelStudent.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * Student Grade Management API
 * OpenAPI spec version: 1.0
 */
import type { Links } from './links';

export interface EntityModelStudent {
  _links?: Links;
  email?: string;
  firstName?: string;
  id?: number;
  lastName?: string;
  username?: string;
}